requirements.txt
anthropic>=0.40.0
pydantic>=2.0
pandas>=2.0
python-dotenv>=1.0
pytest>=8.0
